Read 5 times from camera to ensure frame is new
Prior to this change, CV would, running on the roboRio
and printing via CVReadAndPrintCommand, process correctly
and then print the same output for the next couple shots.

This ensures that the frame being used by hsvThresholding
is always new. Shrinking the camera's buffer with
opencv does not work due to v4l issues.